backend/requirements.txt
alembic==1.11.1
aramgg-poro==0.0.7
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.17
typing_extensions==4.7.0
urllib3==2.0.3
Werkzeug==2.3.6
